Update index.html to use search box text for query
This change fixes "reloadSnippets" to use the text inside the search box
to form a query, and it corrects the event listener from "onkeydown" to
"onchange". This fixes a couple issues with "onkeydown". First, the
searchBox value only gets its new value (as a result of the key press)
after the onkeydown function returns. Second, onkeydown doesn't address
other ways of changing the text (such as pasting using a mouse).