Renames this project "search-server"
The original project name was "gutendemo" which was supposed to reflect
the fact that this project would index and return snippets from text
from Project Gutenberg. Instead, the focus will be on returning generic
documents which contain a title and a body (with potential other fields
as an optional extension for the user). These documents would be about
the size of a reddit comment.