Adds offline indexer using tantivy
The indexer assumes documents are stored in .bz2 files in the "corpus"
directory. The .bz2 is assumed to contain 1 file, which is json
formatted with one object per line, where each object has an "author"
field and a "body" field.

This format matches a common dump of reddit comments available online.

This functionality will be brought into the server in forthcoming
commits.